Fix: use int rather than off_t for offsets
off_t can be a 64-bit type on 32-bit systems, leading to issues on
32-bit big endian systems (ppc32) where the inline assembler expects a
32-bit register.

There is no point in using off_t here.

+++ include/rseq/rseq-skip.h
@@ -13,7 +13,7 @@ int rseq_cmpeqv_storev(intptr_t *v, intptr_t expect, intptr_t newv, int cpu)
 
 static inline __attribute__((always_inline))
 int rseq_cmpnev_storeoffp_load(intptr_t *v, intptr_t expectnot,
-			       off_t voffp, intptr_t *load, int cpu)
+			       int voffp, intptr_t *load, int cpu)
 {
 	return -1;
 }
@@ -65,7 +65,7 @@ int rseq_cmpeqv_trymemcpy_storev_release(intptr_t *v, intptr_t expect,
 }
 
 static inline __attribute__((always_inline))
-int rseq_deref_loadoffp(void *p, off_t voffp, intptr_t *load, int cpu)
+int rseq_deref_loadoffp(void *p, int voffp, intptr_t *load, int cpu)
 {
 	return -1;
 }
